requirements.txt
flask
redis
pytest
requests
docker
mysqlclient==1.4.6
Flask-MySQLdb==0.2.0

